% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/area.study_area_circle.R
\name{area.study_area_circle}
\alias{area.study_area_circle}
\title{area.study_area_circle}
\usage{
area.study_area_circle(xCenter, yCenter, radius, name, background.grid,
  cellsize_background, Nx_background, Ny_background)
}
\arguments{
\item{xCenter}{}

\item{yCenter}{}

\item{radius}{}

\item{name}{}

\item{background.grid}{}

\item{cellsize_background}{}

\item{Nx_background}{}

\item{Ny_background}{}
}
\description{
study_area_circle
}
\examples{
area.study_area_circle()
}
\keyword{area}
\keyword{study}

