% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/area.study_area_norway.R
\name{area.study_area_norway}
\alias{area.study_area_norway}
\title{area.study_area_norway}
\usage{
area.study_area_norway(shp, name, background.grid, cellsize_background)
}
\arguments{
\item{shp}{shapefile}

\item{name}{}

\item{background.grid}{}

\item{cellsize_background}{}
}
\description{
study_area_norway
}
\examples{
area.study_area_norway()
}
\keyword{area}
\keyword{study}

