% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dem.covariates.R
\name{dem.covariates}
\alias{dem.covariates}
\title{dem covariates}
\usage{
dem.covariates(dem.grid, Nx_dem, Ny_dem, cellsize_dem, coverage.grid,
  cellsize_coverage, UTM, border, FIGURE = TRUE, path)
}
\arguments{
\item{dem.grid,}{}

\item{Nx_dem,}{}

\item{Ny_dem,}{}

\item{cellsize_dem,}{}

\item{coverage.grid,}{}

\item{cellsize_coverage,}{}

\item{UTM,}{}

\item{border,}{}

\item{path}{}

\item{FIGURE=TRUE,}{}
}
\description{
dem covariates
}
\examples{
dem.covariates()
}
\keyword{dem}

