% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geography.covariates.R
\name{geography.covariates}
\alias{geography.covariates}
\title{geography.covariates}
\usage{
geography.covariates(dem.grid, Nx_dem, Ny_dem, cellsize_dem, coverage.grid,
  cellsize_coverage, UTM, border, coordinates_coast, coordinates_border,
  FIGURE = TRUE, path)
}
\arguments{
\item{dem.grid,}{}

\item{Nx_dem,}{}

\item{Ny_dem,}{}

\item{cellsize_dem,}{}

\item{coverage.grid,}{}

\item{cellsize_coverage,}{}

\item{UTM,}{}

\item{border,}{}

\item{coordinates_coast,}{}

\item{coordinates_border,}{}

\item{path}{}

\item{FIGURE=TRUE,}{}
}
\description{
geography covariates
}
\examples{
geography.covariates()
}
\keyword{geography}

