% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rr.correction.foerland.R
\name{rr.correction.foerland}
\alias{rr.correction.foerland}
\title{rr.correction.foerland}
\usage{
rr.correction.foerland(RR_values, RR_metadata, RR_covariate, TAMRR_values,
  TAMRR_metadata, TAMRR_covariate, month = month, CV = FALSE)
}
\arguments{
\item{RR_values}{}

\item{RR_metadata}{}

\item{RR_covariate}{}

\item{TAMRR_values}{}

\item{TAMRR_metadata}{}

\item{TAMRR_covariate}{}

\item{month}{}

\item{CV}{cross validation default FALSE}
}
\description{
daily RR correction. Based on DNMI KLIMA, Manual for operational correction of Nordic precipitation data, 24/96
P_cor<-exposition_coef*(P_mes+mean daily evaporation loss+wetting loss). All the constants come from the report
}
\examples{
rr.correction.foerland()
}
\keyword{RR}
\keyword{correction}

