% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/area.study_area_square.R
\name{study_area_square}
\alias{study_area_square}
\title{study_area_square}
\usage{
study_area_square(xMin, xMax, yMin, yMax, name, background.grid,
  cellsize_background)
}
\arguments{
\item{xMin}{}

\item{xMax}{}

\item{yMin}{}

\item{yMax}{}

\item{name}{}

\item{background.grid}{}

\item{cellsize_background}{}
}
\description{
study_area_square
}
\examples{
study_area_square()
}
\keyword{area}
\keyword{study}

