% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tam.interpolation.point.R
\name{tam.interpolation.point}
\alias{tam.interpolation.point}
\title{tam.interpolation.point}
\usage{
tam.interpolation.point(TAM_metadata, TAM_values, TAM_covariate,
  target_metadata, target_covariate, month, CV, method = "SeNorge")
}
\arguments{
\item{TAM_metadata}{}

\item{TAM_values}{}

\item{TAM_covariate}{}

\item{target_metadata}{}

\item{target_covariate}{}

\item{month}{}

\item{CV}{}

\item{method}{}
}
\description{
tam.interpolation.point
}
\examples{
tam.interpolation.point()
}
\keyword{TAM}

