% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tam.interpolation.point.senorge.R
\name{tam.interpolation.point.senorge}
\alias{tam.interpolation.point.senorge}
\title{tam.interpolation.point.senorge}
\usage{
tam.interpolation.point.senorge(TAM_metadata, TAM_values, TAM_covariate,
  target_metadata, target_covariate, month, CV)
}
\arguments{
\item{TAM_metadata}{}

\item{TAM_values}{}

\item{TAM_covariate}{}

\item{target_metadata}{}

\item{target_covariate}{}

\item{month}{}

\item{CV}{}
}
\description{
TAM SeNorge Interpolation method v1.1
 Tveito, O. E., E. Førland, R. heino, I. Hansen-Bauer, H. Alexandersson, B. Dahlström, A.
          Drebs, C. Kern-Hansen, T. Jónsson, E. Vaarby Laursen and Y. Westman, 2000: “Nordic
          temperature maps”, DNMI Report no. 09/00
 Tveito, O. E., I. Bjørdal, A. O. Skjelvåg and B. Aune, 2005: ”A GIS-based agro-ecological
           decision system based on gridded climatology, Met. Apps., Vol. 12: 1, p. 57-68
}
\details{
Synthesis note:
 Mohr M.  New Routines for Gridding of Temperature and Precipitation Observations for
           “seNorge.no”, MET report No. 08/2008
}
\examples{
tam.interpolation.point.senorge()
}
\keyword{TAM}

